% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alvc}
\alias{alvc}
\title{Set of error estimates (variance-covariance matrix) for the relative pollen productivity estimates for the Gaussian plume model.}
\format{
A matrix with identical columns and rows. The diagonal represents error estimates. Other values are covariances, here empty (data not available).
}
\source{
Abraham, V., Oušková, V., & Kuneš, P. 2014. Present-day vegetation helps quantifying past land cover in selected regions of the Czech Republic. PLoS ONE 9: e100117.
}
\usage{
alvc
}
\description{
Set of error estimates (variance-covariance matrix) for the relative pollen productivity estimates for the Gaussian plume model.
}
\examples{
\dontrun{
 alvc
}
}
\keyword{datasets}
