% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg}
\alias{avg}
\title{Set of relative pollen productivity estimates for the Gaussian plume model}
\format{
A data frame with columns:
\describe{
 \item{alpha}{Relative pollen productivity to Poaceae = 1.}
 \item{vg}{Fall speed of pollen in m/s.}
}
}
\source{
Kuneš, P., Abraham, V., Werchan, B., Plesková, Z., Fajmon, K., Jamrichová, E., & Roleček, J. 2019. Relative pollen productivity estimates for vegetation reconstruction in central-eastern Europe inferred at local and regional scales. The Holocene 29: 1708–1719.
}
\usage{
avg
}
\description{
Set of relative pollen productivity estimates for the Gaussian plume model
}
\examples{
\dontrun{
 avg
}
}
\keyword{datasets}
