% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPf.R
\name{KPf}
\alias{KPf}
\title{Pollen dispersal-deposition coefficient}
\usage{
KPf(vg, u, Zmax, radius, model, dwm)
}
\arguments{
\item{vg}{fall speed of pollen in m/s}

\item{u}{wind speed in m/s}

\item{Zmax}{maximum extent of the region in metres}

\item{model}{type of the deposition model for the type of sedimentation basin: "peatland" (Prentice, 1985), "lake" (Sugita, 1993)}

\item{dwm}{type of dispersal model used: "gpm neutral" (Prentice, 1985), "lsm unstable" - from \code{\link{DispersalFactorK}} function in \code{DISQOVER} package (Theuerkauf, et al. 2016)}
}
\description{
Calculates the taxon-specific pollen dispersal-deposition coefficient required for \code{\link{REVEALS}} function.
}
\examples{


}
\references{
Prentice, I.C. 1985. Pollen representation, source area, and basin size: Toward a unified theory of pollen analysis. Quaternary Research 23: 76–86.

Sugita, S. 1993. A Model of Pollen Source Area for an Entire Lake Surface. Quaternary Research 39: 239–244.

Sugita, S. 2007. Theory of quantitative reconstruction of vegetation I: pollen from large sites REVEALS regional vegetation composition. Holocene 17: 229–241.

Theuerkauf, M., Couwenberg, J., Kuparinen, A., & Liebscher, V. 2016. A matter of dispersal: REVEALSinR introduces state-of-the-art dispersal models to quantitative vegetation reconstruction. Vegetation History and Archaeobotany 25: 541–553.
}
