% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC_Rybarenska}
\alias{PC_Rybarenska}
\title{Pollen counts from Rybárenská slať in Šumava Mts., Czech Republic}
\format{
A data frames with columns as time windows and rows:
\describe{
 \item{Radius}{Site radius in m.}
 \item{Model}{model type: 1 = Prentice bog model, 2 = Sugita lake model.}
 \item{Abies...}{Pollen counts for each taxon.}
}
}
\source{
Carter, V.A., Chiverrell, R.C., Clear, J.L., Kuosmanen, N., Moravcová, A., Svoboda, M., Svobodová-Svitavská, H., Leeuwen, V., Van Leeuwen, J., van der Knaap, W.O., & Kuneš, P. 2018. Quantitative palynology informing conservation ecology in the Bohemian/Bavarian Forests of Central Europe. Frontiers in Plant Science 8: 1–14.

Svobodová, H., Soukupová, L., & Reille, M. 2002. Diversified development of mountain mires, Bohemian Forest, Central Europe, in the last 13,000 years. Quaternary International 91: 123–135.
}
\usage{
PC_Rybarenska
}
\description{
Pollen counts from Rybárenská slať in Šumava Mts., Czech Republic
}
\examples{
\dontrun{
 PC_Rybarenska
}
}
\keyword{datasets}
