% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REVEALS.R
\name{REVEALS}
\alias{REVEALS}
\title{REVEALS (Regional Vegetation Reconstruction from Large Sites)}
\usage{
REVEALS(pollen_counts, avg, alvc, u, Zmax, dwm)
}
\arguments{
\item{pollen_counts}{contains EITHER data frame of a single pollen count spreadsheet with the following structure:
first row includes ages or depths or sample names (while using multiple sites colnames have to match at all sites)
second row includes radius in metres
third row includes model type: 1 = Prentice bog model, 2 = Sugita lake model
fourth row and onwards include taxa with pollen counts
OR list of data frames of multiple sites with the structure described above}

\item{avg}{contains data frame where rows represent all calculated taxa, the first column is taxon name, the second column is alpha (pollen productivity estimate) and the third row is vg (fall speed of pollen)}

\item{alvc}{contains data frame with variances and covariances of PPEs for all taxa arranged in columns and rows of the same size}

\item{u}{wind speed in m/s}

\item{Zmax}{– maximum extent of the region in metres}

\item{dwm}{type of dispersal model used: "gpm neutral", "lsm unstable"}
}
\value{
Returns a list containing the following objects: \code{V_sites} – vegetation estimates for particular sites;
\code{SE_sites} – standard errors of vegetation estimantes for particular sites;
\code{Mean_V} – mean vegetation estimates for all sites;
\code{Mean_SE} – standard errors of mean vegetation estimates for all sites;
\code{No_sites} – number of sites used for estimating particular time window
}
\description{
Calculates REVEALS vegetation estimates (Sugita, 2007) from a matrix of pollen counts or a list of pollen count matrices (multiple sites) in \code{pollen_counts}.
The function enables using deposition models for bogs and lakes, and two dispersal models (Gaussian plume model, Lagrangian stochastic model taken from the DISQOVER package (Theuerkauf,et al. 2016)).
}
\note{
Citation of this programme: Abraham, V., Oušková, V., & Kuneš, P. 2014. Present-day vegetation helps quantifying past land cover in selected regions of the Czech Republic. PLoS ONE 9: e100117.
}
\examples{
# RUN REVEALS WITH MULTIPLE SITES
REVEALS.mult.sites <- REVEALS(list(PC_Cerne, PC_Prasilske, PC_Rybarenska), avg, alvc, 3, 100000, "gpm neutral")




}
\references{
Abraham, V., Oušková, V., & Kuneš, P. 2014. Present-day vegetation helps quantifying past land cover in selected regions of the Czech Republic. PLoS ONE 9: e100117.

Sugita, S. 2007. Theory of quantitative reconstruction of vegetation I: pollen from large sites REVEALS regional vegetation composition. Holocene 17: 229–241.

Theuerkauf, M., Couwenberg, J., Kuparinen, A., & Liebscher, V. 2016. A matter of dispersal: REVEALSinR introduces state-of-the-art dispersal models to quantitative vegetation reconstruction. Vegetation History and Archaeobotany 25: 541–553.
}
