% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alvc}
\alias{alvc}
\title{Set of error estimates (variance-covariance matrix) for the relative pollen productivity estimates for the Gaussian plume model.}
\format{
A matrix with identical columns and rows. The diagonal represents error estimates. Other values are covariances, here empty (data not available).
}
\source{
Kuneš, P., Abraham, V., Werchan, B., Plesková, Z., Fajmon, K., Jamrichová, E., & Roleček, J. 2019. Relative pollen productivity estimates for vegetation reconstruction in central-eastern Europe inferred at local and regional scales. The Holocene 29: 1708–1719.
}
\usage{
alvc
}
\description{
Set of error estimates (variance-covariance matrix) for the relative pollen productivity estimates for the Gaussian plume model.
}
\examples{
\dontrun{
 alvc
}
}
\keyword{datasets}
