% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg}
\alias{avg}
\title{Set of relative pollen productivity estimates for the Gaussian plume model}
\format{
A data frame with columns:
\describe{
 \item{alpha}{Relative pollen productivity to Poaceae = 1.}
 \item{vg}{Fall speed of pollen in m/s.}
}
}
\source{
Abraham, V., Oušková, V., & Kuneš, P. 2014. Present-day vegetation helps quantifying past land cover in selected regions of the Czech Republic. PLoS ONE 9: e100117.
}
\usage{
avg
}
\description{
Set of relative pollen productivity estimates for the Gaussian plume model
}
\examples{
\dontrun{
 avg
}
}
\keyword{datasets}
